#27 - moves operation tests to module
