+++ feature/operations/src/androidTest/java/br/com/rbrthmn/operations/AccountsDropdownMenuUITest.kt
@@ -18,7 +18,7 @@
  *
  */
 
-package br.com.rbrthmn.ui.financialcompanion.components
+package br.com.rbrthmn.operations
 
 import androidx.activity.ComponentActivity
 import androidx.compose.ui.test.assertCountEquals
@@ -31,11 +31,15 @@ import androidx.compose.ui.test.onLast
 import androidx.compose.ui.test.onNodeWithContentDescription
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performClick
-import br.com.rbrthmn.R
-import br.com.rbrthmn.operations.OperationOriginAccount
+import br.com.rbrthmn.operations.components.ACCOUNTS_DROPDOWN_MENU_ITEM_TAG
+import br.com.rbrthmn.operations.components.ADD_SIMPLE_ACCOUNT_DIALOG_TAG
+import br.com.rbrthmn.operations.components.AccountsDropdownMenu
+import br.com.rbrthmn.operations.components.OperationOriginAccount
+import br.com.rbrthmn.ui.BaseUITest
+import br.com.rbrthmn.ui.R
 import org.junit.Test
 
-class AccountsDropdownMenuUITest : br.com.rbrthmn.ui.BaseUITest() {
+class AccountsDropdownMenuUITest : BaseUITest() {
     override val composeTestRule = createAndroidComposeRule<ComponentActivity>()
 
     override fun setup() = composeTestRule.setContent {
