+++ feature/operations/src/androidTest/java/br/com/rbrthmn/operations/OperationsListUITest.kt
@@ -1,24 +1,4 @@
-/*
- *
- * Copyright (C) 2022 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Modifications made by Roberto Kenzo Hamano, 2024
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package br.com.rbrthmn.ui.financialcompanion.components
+package br.com.rbrthmn.operations
 
 import androidx.activity.ComponentActivity
 import androidx.compose.ui.test.assertIsDisplayed
@@ -32,16 +12,15 @@ import androidx.compose.ui.test.onNodeWithContentDescription
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performTextInput
-import br.com.rbrthmn.R
-import br.com.rbrthmn.operations.OperationsScreenViewModel
-import br.com.rbrthmn.ui.financialcompanion.screens.operations.components.ACCOUNTS_DROPDOWN_ICON_TAG
-import br.com.rbrthmn.ui.financialcompanion.screens.operations.components.ACCOUNTS_DROPDOWN_MENU_TAG
-import br.com.rbrthmn.ui.financialcompanion.screens.operations.components.NEW_OPERATION_DIALOG_TAG
-import br.com.rbrthmn.ui.financialcompanion.screens.operations.components.OPERATION_TYPES_DROPDOWN_MENU_TAG
-import br.com.rbrthmn.ui.financialcompanion.screens.operations.components.OperationsListCard
+import br.com.rbrthmn.operations.components.ACCOUNTS_DROPDOWN_ICON_TAG
+import br.com.rbrthmn.operations.components.ACCOUNTS_DROPDOWN_MENU_TAG
+import br.com.rbrthmn.operations.components.NEW_OPERATION_DIALOG_TAG
+import br.com.rbrthmn.operations.components.OPERATION_TYPES_DROPDOWN_MENU_TAG
+import br.com.rbrthmn.operations.components.OperationsListCard
 import br.com.rbrthmn.ui.onNodeWithStringId
 import br.com.rbrthmn.ui.utils.ResourceStringProvider
 import org.junit.Test
+import br.com.rbrthmn.ui.R as uiR
 
 class OperationsListUITest : br.com.rbrthmn.ui.BaseUITest() {
     override val composeTestRule = createAndroidComposeRule<ComponentActivity>()
@@ -68,7 +47,7 @@ class OperationsListUITest : br.com.rbrthmn.ui.BaseUITest() {
             .onNodeWithStringId(R.string.add_operation_button)
             .performClick()
         composeTestRule
-            .onNodeWithStringId(R.string.cancel_button)
+            .onNodeWithStringId(uiR.string.cancel_button)
             .performClick()
         val dialog = composeTestRule.onNodeWithTag(NEW_OPERATION_DIALOG_TAG)
 
@@ -81,7 +60,7 @@ class OperationsListUITest : br.com.rbrthmn.ui.BaseUITest() {
             .onNodeWithStringId(R.string.add_operation_button)
             .performClick()
         composeTestRule
-            .onNodeWithStringId(R.string.save_button)
+            .onNodeWithStringId(uiR.string.save_button)
             .performClick()
         val dialog = composeTestRule.onNodeWithTag(NEW_OPERATION_DIALOG_TAG)
 
@@ -96,7 +75,7 @@ class OperationsListUITest : br.com.rbrthmn.ui.BaseUITest() {
                 VALID_DESCRIPTION
             )
             onNodeWithStringId(R.string.operation_value_hint).performTextInput(VALID_VALUE)
-            onNodeWithContentDescription(activity.getString(R.string.drop_down_arrow_icon_description)).performClick()
+            onNodeWithContentDescription(activity.getString(uiR.string.drop_down_arrow_icon_description)).performClick()
             onNodeWithTag(OPERATION_TYPES_DROPDOWN_MENU_TAG)
                 .onChildren()
                 .onFirst()
@@ -117,7 +96,7 @@ class OperationsListUITest : br.com.rbrthmn.ui.BaseUITest() {
                 .onChildren()
                 .onFirst()
                 .performClick()
-            onNodeWithStringId(R.string.save_button).performClick()
+            onNodeWithStringId(uiR.string.save_button).performClick()
             val dialog = onNodeWithTag(NEW_OPERATION_DIALOG_TAG)
 
             dialog.assertIsNotDisplayed()
