+++ feature/operations/src/test/java/br/com/rbrthmn/operations/OperationsScreenViewModel.kt
@@ -18,13 +18,11 @@
  *
  */
 
-package rbthmn.viewmodels
+package br.com.rbrthmn.operations
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
-import br.com.rbrthmn.operations.OperationType
 import br.com.rbrthmn.operations.OperationsScreenContract.Intent
-import br.com.rbrthmn.operations.OperationsScreenViewModel
 import br.com.rbrthmn.operations.OperationsScreenViewModel.Companion.TOTAL_BALANCE_MOCK
 import br.com.rbrthmn.operations.OperationsScreenViewModel.Companion.TOTAL_INCOME_MOCK
 import br.com.rbrthmn.operations.OperationsScreenViewModel.Companion.TOTAL_OUTCOME_MOCK
